Stop using middleman-cloudfront
Cloudfront starts charging you once you have invalidated more than 1000
paths. The middleman-cloudfront extension in turn is a bit dumb, adding
paths to invalidations for newly added files. Even if it was smarter,
changing 50 files would still consume 50 invalidation paths.

To work around this we just use the AWS CLI, which allows creating of
invalidations using wildcards. This means that any push will invalidate
the entire cache, which is a bit of a waste, but it also saves me money,
which is more important at this time.