Fix wrapping of long lines in pre blocks
As mentioned in this Reddit comment [1], the release post looked a bit
odd on mobile devices. This was due to a few very long lines in code
blocks not getting wrapped properly. Apparently the use of CSS grids
means overflowing is handled differently, and to fix that the grid items
must have a minimum width of zero [2]. Because that makes total sense.

[1]: https://www.reddit.com/r/inko/comments/13lzqdx/inko_0110_released_including_a_native_code/jkumajx/
[2]: https://stackoverflow.com/questions/43311943/prevent-content-from-expanding-grid-items