Add a rudimentary package index
This index is built from a basic YAML file, and uses the GitHub API to
pull in the necessary details such as tags and stars. At some point in
the future we'll likely outgrow this setup, at which point we'll need a
dedicated service, but for the time being this should be sufficient.