Run the deploy job directly when updating packages
It seems that running the deploy as a separate job results in the
website not updating. My guess is that even though the checked out ref
is "main", somehow the package updates aren't included.

To work around this, we just call `rake deploy` in the same job that
updated the package data. This should hopefully fix things.